+++ protobuf/identitypb/identity_test.go
@@ -58,7 +58,6 @@ func TestIdentity_Validate(t *testing.T) {
 		}
 
 		for _, c := range cases {
-			c := c // capture loop variable
 			t.Run(c.Desc, func(t *testing.T) {
 				t.Parallel()
 
@@ -160,7 +159,6 @@ func TestIdentity_ParseAndMustParse(t *testing.T) {
 		}
 
 		for _, c := range cases {
-			c := c // capture loop variable
 			t.Run(c.Desc, func(t *testing.T) {
 				t.Parallel()